Compatilibilty for web and react-native
- Implemented own util.inherits function
- Moved functions which need the `fs` module to separate file
