+++ encoding/unicode.js
@@ -1,5 +1,5 @@
 'use strict';
-var util = require('util'),
+var util = require('../util'),
   Match = require ('../match');
 
 /**
