+++ index.js
@@ -1,6 +1,3 @@
-
-var fs = require('fs');
-
 var utf8  = require('./encoding/utf8'),
   unicode = require('./encoding/unicode'),
   mbcs    = require('./encoding/mbcs'),
@@ -76,49 +73,6 @@ module.exports.detect = function(buffer, opts) {
   else {
     return matches.length > 0 ? matches[0].name : null;
   }
-};
-
-module.exports.detectFile = function(filepath, opts, cb) {
-  if (typeof opts === 'function') {
-    cb = opts;
-    opts = undefined;
-  }
-
-  var fd;
-
-  var handler = function(err, buffer) {
-    if (fd) {
-      fs.closeSync(fd);
-    }
-
-    if (err) return cb(err, null);
-    cb(null, self.detect(buffer, opts));
-  };
-
-  if (opts && opts.sampleSize) {
-    fd = fs.openSync(filepath, 'r'),
-      sample = Buffer.allocUnsafe(opts.sampleSize);
-
-    fs.read(fd, sample, 0, opts.sampleSize, null, function(err) {
-      handler(err, sample);
-    });
-    return;
-  }
-
-  fs.readFile(filepath, handler);
-};
-
-module.exports.detectFileSync = function(filepath, opts) {
-  if (opts && opts.sampleSize) {
-    var fd = fs.openSync(filepath, 'r'),
-      sample = Buffer.allocUnsafe(opts.sampleSize);
-
-    fs.readSync(fd, sample, 0, opts.sampleSize);
-    fs.closeSync(fd);
-    return self.detect(sample, opts);
-  }
-
-  return self.detect(fs.readFileSync(filepath), opts);
 };
 
 // Wrappers for the previous functions to return all encodings
